Add tests for handlers. Add Epic and Subtask TypeAdapters. Rename packages.
