+++ src/handlers/EpicHandler.java
@@ -7,8 +7,8 @@ import managers.TaskManager;
 import tasks.Epic;
 import tasks.Subtask;
 import tasks.Task;
-import type_adapters.EpicAdapter;
-import type_adapters.SubtaskAdapter;
+import type.adapters.EpicAdapter;
+import type.adapters.SubtaskAdapter;
 
 import java.io.IOException;
 import java.io.InputStream;
