+++ src/handlers/PrioritizedHandler.java
@@ -5,8 +5,8 @@ import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
 import managers.TaskManager;
 import tasks.Task;
-import type_adapters.DurationAdapter;
-import type_adapters.LocalDateTimeAdapter;
+import type.adapters.DurationAdapter;
+import type.adapters.LocalDateTimeAdapter;
 
 import java.io.IOException;
 import java.time.Duration;
