+++ test/handlers/SubtaskHandlerTest.java
@@ -12,7 +12,7 @@ import tasks.Epic;
 import tasks.Subtask;
 import tasks.Task;
 import tasks.TaskStatus;
-import type_adapters.SubtaskAdapter;
+import type.adapters.SubtaskAdapter;
 
 import java.io.IOException;
 import java.net.URI;
